+++ packages/api/datewise/lib/plain-date/past-infinity-date.ts
@@ -8,6 +8,7 @@ import { PlainDateObject } from './plain-date-object.js'
 import { PlainDate, PlainDateInput } from './plain-date.js'
 import { factory } from './plain-date.factory.js'
 import { IsoWeekParity } from './iso-week-parity.js'
+import { IsoWeekday } from '#src/common/iso-weekday.js'
 
 export class PastInfinityDate implements PlainDate {
   isSame (otherDate: PlainDateInput, _unit?: PlainDateUnit): boolean {
@@ -194,9 +195,32 @@ export class PastInfinityDate implements PlainDate {
     throw new Error('cannot access day of week of -infinity')
   }
 
-  isoWeekday (): number {
+  isoWeekday (): IsoWeekday
+  isoWeekday (day: IsoWeekday): this
+  isoWeekday (day?: IsoWeekday): IsoWeekday | this {
+    if (day !== undefined) {
+      return this
+    }
+
     throw new Error('cannot access ISO weekday of -infinity')
   }
+
+  nextIsoWeekday (_day: IsoWeekday): this {
+    return this
+  }
+
+  nextOrSameIsoWeekday (_day: IsoWeekday): this {
+    return this
+  }
+
+  previousIsoWeekday (_day: IsoWeekday): this {
+    return this
+  }
+
+  previousOrSameIsoWeekday (_day: IsoWeekday): this {
+    return this
+  }
+
 
   isoWeek (): number {
     throw new Error('cannot access ISO week of -infinity')